requirements.txt
torch
torchvision
matplotlib
numpy
scikit-learn
tqdm
pyyaml
opencv-python
streamlit
seaborn
